# api/openapi.yaml
openapi: 3.0.3
info:
  title: Miri Autonomous Agent API
  description: API for interacting with the Miri autonomous agent service.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server

components:
  securitySchemes:
    ServerKey:
      type: apiKey
      in: header
      name: X-Server-Key
    BasicAuth:
      type: http
      scheme: basic

  schemas:
    Session:
      type: object
      properties:
        id:
          type: string
        soul:
          type: string
        total_tokens:
          type: integer
        prompt_tokens:
          type: integer
        output_tokens:
          type: integer
        total_cost:
          type: number
          format: float
    Message:
      type: object
      properties:
        prompt:
          type: string
        response:
          type: string
    Usage:
      type: object
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer
        total_cost:
          type: number
          format: float
    HumanInfo:
      type: object
      properties:
        id:
          type: string
        data:
          type: object
          additionalProperties:
            type: string
        notes:
          type: string
    Config:
      type: object
      properties:
        storage_dir:
          type: string
        server:
          type: object
          properties:
            addr:
              type: string
            key:
              type: string
            admin_user:
              type: string
              default: admin
            admin_pass:
              type: string
              default: admin-password
        miri:
          $ref: '#/components/schemas/MiriSubConfig'
        models:
          type: object
          properties:
            mode:
              type: string
            providers:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/ProviderConfig'
        agents:
          type: object
          properties:
            debug:
              type: boolean
            subagents:
              type: integer
            defaults:
              type: object
              properties:
                model:
                  type: object
                  properties:
                    primary:
                      type: string
                    fallbacks:
                      type: array
                      items:
                        type: string
        channels:
          type: object
          properties:
            whatsapp:
              type: object
              properties:
                enabled:
                  type: boolean
                allowlist:
                  type: array
                  items:
                    type: string
                blocklist:
                  type: array
                  items:
                    type: string
            irc:
              type: object
              properties:
                enabled:
                  type: boolean
                host:
                  type: string
                port:
                  type: integer
                tls:
                  type: boolean
                nick:
                  type: string
                user:
                  type: string
                realname:
                  type: string
                channels:
                  type: array
                  items:
                    type: string
                password:
                  type: string
                nickserv:
                  type: object
                  properties:
                    enabled:
                      type: boolean
                    password:
                      type: string
                allowlist:
                  type: array
                  items:
                    type: string
                blocklist:
                  type: array
                  items:
                    type: string

    MiriSubConfig:
      type: object
      properties:
        brain:
          $ref: '#/components/schemas/BrainConfig'

    BrainConfig:
      type: object
      properties:
        embeddings:
          $ref: '#/components/schemas/EmbeddingConfig'

    EmbeddingConfig:
      type: object
      properties:
        use_native_embeddings:
          type: boolean
        model:
          $ref: '#/components/schemas/EmbeddingModelConfig'

    EmbeddingModelConfig:
      type: object
      properties:
        type:
          type: string
        api_key:
          type: string
        model:
          type: string
        url:
          type: string

    ProviderConfig:
      type: object
      properties:
        baseUrl:
          type: string
        apiKey:
          type: string
        api:
          type: string
        models:
          type: array
          items:
            $ref: '#/components/schemas/ModelConfig'
    ModelConfig:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        contextWindow:
          type: integer
        maxTokens:
          type: integer
        reasoning:
          type: boolean
        input:
          type: array
          items:
            type: string
        cost:
          type: object
          properties:
            input:
              type: number
            output:
              type: number
            cacheRead:
              type: number
            cacheWrite:
              type: number

    Task:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        cron_expression:
          type: string
        prompt:
          type: string
        active:
          type: boolean
        needed_skills:
          type: array
          items:
            type: string
        last_run:
          type: string
          format: date-time
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        report_session:
          type: string
        report_channels:
          type: array
          items:
            type: string

    Skill:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        version:
          type: string
        tags:
          type: array
          items:
            type: string
        directory:
          type: string
        full_content:
          type: string

    SkillCommand:
      type: object
      properties:
        name:
          type: string
        description:
          type: string

paths:
  # --- Standard API (X-Server-Key) ---
  /api/v1/prompt:
    post:
      summary: Send a prompt to the agent
      security:
        - ServerKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prompt:
                  type: string
                model:
                  type: string
                temperature:
                  type: number
                max_tokens:
                  type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string

  /api/v1/prompt/stream:
    get:
      summary: Stream a prompt response via SSE
      security:
        - ServerKey: []
      parameters:
        - name: prompt
          in: query
          required: true
          schema:
            type: string
        - name: model
          in: query
          schema:
            type: string
      responses:
        '200':
          description: SSE stream of response chunks
          content:
            text/event-stream:
              schema:
                type: string

  /api/v1/interaction:
    post:
      summary: Manage sessions or check global status
      security:
        - ServerKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [status]
      responses:
        '200':
          description: Successful response

  /ws:
    get:
      summary: WebSocket for interactive streaming
      description: |
        WebSocket endpoint for real-time interaction and background task notifications.
        When a background task completes, a message is sent in the following format:
        ```json
        {
          "type": "task_complete",
          "task_id": "uuid",
          "task_name": "task name",
          "response": "result string"
        }
        ```
      security:
        - ServerKey: []
      parameters:
        - name: channel
          in: query
          schema:
            type: string
        - name: device
          in: query
          schema:
            type: string
      responses:
        '101':
          description: WebSocket upgrade

  /api/v1/files/{filepath}:
    get:
      summary: Download a file from the local storage
      security:
        - ServerKey: []
      parameters:
        - name: filepath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The file content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: File not found
        '403':
          description: Access denied

  # --- Admin API (Basic Auth) ---
  /api/admin/v1/health:
    get:
      summary: Check health of the admin API
      security:
        - BasicAuth: []
      responses:
        '200':
          description: Healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string

  /api/admin/v1/config:
    get:
      summary: Get current configuration
      security:
        - BasicAuth: []
      responses:
        '200':
          description: Current config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Config'
    post:
      summary: Update configuration
      security:
        - BasicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Config'
      responses:
        '200':
          description: Config updated

  /api/admin/v1/human:
    get:
      summary: List all stored human information
      security:
        - BasicAuth: []
      responses:
        '200':
          description: List of human info
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HumanInfo'
    post:
      summary: Store human information
      security:
        - BasicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HumanInfo'
      responses:
        '200':
          description: Human info saved

  /api/admin/v1/channels:
    post:
      summary: Perform actions on communication channels
      security:
        - BasicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                channel:
                  type: string
                action:
                  type: string
                  enum: [status, enroll, send, devices, chat]
                device:
                  type: string
                message:
                  type: string
                prompt:
                  type: string
      responses:
        '200':
          description: Action performed

  /api/admin/v1/sessions:
    get:
      summary: List active session IDs
      security:
        - BasicAuth: []
      responses:
        '200':
          description: List of session IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /api/admin/v1/sessions/{id}:
    get:
      summary: Get session details
      security:
        - BasicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'

  /api/admin/v1/sessions/{id}/stats:
    get:
      summary: Get session token and cost statistics
      security:
        - BasicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
                  total_tokens:
                    type: integer
                  prompt_tokens:
                    type: integer
                  output_tokens:
                    type: integer
                  total_cost:
                    type: number
                    format: float

  /api/admin/v1/sessions/{id}/history:
    get:
      summary: Get session message history
      security:
        - BasicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session history
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                  total_tokens:
                    type: integer

  /api/admin/v1/sessions/{id}/skills:
    get:
      summary: Get loaded skills for a session
      security:
        - BasicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of loaded skill names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /api/admin/v1/skills:
    get:
      summary: List all installed skills
      security:
        - BasicAuth: []
      responses:
        '200':
          description: List of skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Skill'

  /api/admin/v1/skills/commands:
    get:
      summary: List all available agent commands (tools)
      security:
        - BasicAuth: []
      responses:
        '200':
          description: List of available commands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SkillCommand'

  /api/admin/v1/skills/{name}:
    get:
      summary: Get skill details
      security:
        - BasicAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Skill details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
    delete:
      summary: Remove a skill
      security:
        - BasicAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Skill removed


  /api/admin/v1/tasks:
    get:
      summary: List all recurring tasks
      security:
        - BasicAuth: []
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'

  /api/admin/v1/tasks/{id}:
    get:
      summary: Get task details
      security:
        - BasicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'